rename BlogSidebar -> BlogCategoriesList and styles
