+++ src/mdxGlobalComponents.js
@@ -7,12 +7,12 @@ import { AnchorScrollNavbar } from './components/AnchorScrollNavbar'
 import { ArrayCTA } from './components/ArrayCTA'
 import { BasicHedgehogImage } from './components/BasicHedgehogImage'
 import { BlogAuthor } from './components/Blog/BlogAuthor'
+import { BlogCategoriesList } from './components/Blog/BlogCategoriesList'
 import { BlogFeaturedImage } from './components/Blog/BlogFeaturedImage'
 import { BlogIntro } from './components/Blog/BlogIntro'
 import { BlogPostLayout } from './components/Blog/BlogPostLayout'
 import { BlogPosts } from './components/Blog/BlogPosts'
 import { BlogShareButtons } from './components/Blog/BlogShareButtons'
-import { BlogSidebar } from './components/Blog/BlogSidebar'
 import { BlogFooter } from './components/BlogFooter'
 import { CallToAction } from './components/CallToAction'
 import { Benefits } from './components/Careers/Benefits'
@@ -103,12 +103,12 @@ export const shortcodes = {
     ArrayCTA,
     BasicHedgehogImage,
     BlogAuthor,
+    BlogCategoriesList,
     BlogFeaturedImage,
     BlogIntro,
     BlogPostLayout,
     BlogPosts,
     BlogShareButtons,
-    BlogSidebar,
     BlogFooter,
     CallToAction,
     Benefits,
